+++ examples/single.js
@@ -11,8 +11,19 @@ function handleChange(value) {
 var Test = React.createClass({
   getInitialState() {
     return {
-      destroy: false
+      destroy: false,
+      value: 'lucy'
     }
+  },
+
+  onChange(e){
+    let value;
+    if (e.target) {
+      value = e.target.value;
+    } else {
+      value = e;
+    }
+    this.setState({value});
   },
 
   handleDestroy() {
@@ -28,12 +39,13 @@ var Test = React.createClass({
     return (
       <div style={{margin: 20}}>
         <h2>Single Select</h2>
+
         <div style={{width: 300}}>
-          <Select value="lucy"
-            dropdownMenuStyle={{maxHeight: 200, overflow: 'auto'}}
-            renderDropdownToBody={location.href.indexOf('renderDropdownToBody') !== -1}
-            style={{width: 500}}
-            onChange={handleChange}>
+          <Select value={this.state.value}
+                  dropdownMenuStyle={{maxHeight: 200, overflow: 'auto'}}
+                  renderDropdownToBody={location.href.indexOf('renderDropdownToBody') !== -1}
+                  style={{width: 500}}
+                  onChange={this.onChange}>
             <Option value="jack">
               <b style={{
                 color: 'red'
@@ -42,20 +54,23 @@ var Test = React.createClass({
             <Option value="lucy">lucy</Option>
             <Option value="disabled" disabled>disabled</Option>
             <Option value="yiminghe">yiminghe</Option>
-          {[1, 2, 3, 4, 5, 6, 7, 8, 9].map(function (i) {
-            return <Option key={i}>{i}</Option>
-          })}
+            {[1, 2, 3, 4, 5, 6, 7, 8, 9].map(function (i) {
+              return <Option key={i}>{i}</Option>
+            })}
           </Select>
         </div>
 
         <h2>native select</h2>
-        <select value="lucy"
-          style={{width: 500}}
-          onChange={handleChange}>
+        <select value={this.state.value}
+                style={{width: 500}}
+                onChange={this.onChange}>
           <option value="jack">jack</option>
           <option value="lucy">lucy</option>
           <option value="disabled" disabled>disabled</option>
           <option value="yiminghe">yiminghe</option>
+          {[1, 2, 3, 4, 5, 6, 7, 8, 9].map(function (i) {
+            return <option value={i} key={i}>{i}</option>
+          })}
         </select>
 
         <p>